// src/app/Components/Models/grafica.model.ts
export class GraficaModel{
    listMaxFloracion: number[] ;
	listBrotesPosPerd: number[] ;
	listBrotes: number[] ;
	listFrutosPosPerd: number[];
	listFrutos: number[] ;
	listNitrogeno: number[];
	listPotasio: number[];
	listFosforo: number[];
	listFronteraAirTierr: number[];
	listHorizontedeSuelo: number[];
	listYear:String[];
}